+++ src/Provider/Provider.js
@@ -7,22 +7,20 @@ const Server = require('../Utils/Server')
 const Request = require('../Utils/Request')
 const Platform = require('../Utils/Platform')
 const Auth = require('../Utils/Auth')
-const Database = require('../Utils/Database')
+const Mongodb = require('../Utils/Database')
 
 const GradeService = require('./Services/Grade')
 
 const url = require('url')
-const validator = require('validator')
+const _path = require('path')
+const jwt = require('jsonwebtoken')
 const winston = require('winston')
-const mongoose = require('mongoose')
-mongoose.set('useCreateIndex', true)
-const Schema = mongoose.Schema
 
 const provAuthDebug = require('debug')('provider:auth')
 const provMainDebug = require('debug')('provider:main')
 
 /**
- * @descripttion Exposes methods for easy manipualtion of the LTI 1.3 standard as a LTI Provider and a "server" object to manipulate the Express instance
+ * @descripttion Exposes methods for easy manipulation of the LTI 1.3 standard as a LTI Provider and a "server" object to manipulate the Express instance
  */
 class Provider {
   // Pre-initiated variables
@@ -46,7 +44,7 @@ class Provider {
     signed: true
   }
 
-  #dbConnection = {}
+  #Database
 
   #connectCallback = () => {}
 
@@ -64,7 +62,8 @@ class Provider {
      * @description Exposes methods for easy manipualtion of the LTI 1.3 standard as a LTI Provider and a "server" object to manipulate the Express instance.
      * @param {String} encryptionkey - Secret used to sign cookies and encrypt other info.
      * @param {Object} database - The Database configurations to open and manage connection, uses MongoDB Driver.
-     * @param {String} database.url - Database Url (Ex: mongodb://localhost/applicationdb).
+     * @param {String} [database.url] - Database Url (Ex: mongodb://localhost/applicationdb).
+     * @param {Object} [database.plugin] - If set, must be the Database object of the desired database plugin.
      * @param {Object} [database.connection] - Database connection options (Ex: user, pass)
      * @param {String} [database.connection.user] - Database user for authentication if needed.
      * @param {String} [database.conenction.pass] - Database pass for authentication if needed.
@@ -83,7 +82,10 @@ class Provider {
   constructor (encryptionkey, database, options) {
     if (options && options.https && (!options.ssl || !options.ssl.key || !options.ssl.cert)) throw new Error('No ssl Key  or Certificate found for local https configuration.')
     if (!encryptionkey) throw new Error('Encryptionkey parameter missing in options.')
-    if (!database || !database.url) throw new Error('Missing database configurations.')
+    if (!database) throw new Error('Missing database configurations.')
+
+    if (!database.plugin) this.#Database = new Mongodb(database)
+    else this.#Database = database.plugin
 
     if (options && options.appUrl) this.#appUrl = options.appUrl
     if (options && options.loginUrl) this.#loginUrl = options.loginUrl
@@ -100,13 +102,15 @@ class Provider {
           winston.format.prettyPrint()
         ),
         transports: [
-          new winston.transports.File({ filename: 'logs/ltijs_error.log',
+          new winston.transports.File({
+            filename: 'logs/ltijs_error.log',
             level: 'error',
             handleExceptions: true,
             maxsize: 250000, // 500kb (with two files)
             maxFiles: 1,
             colorize: false,
-            tailable: true })
+            tailable: true
+          })
         ],
         exitOnError: false
       })
@@ -118,13 +122,15 @@ class Provider {
           winston.format.prettyPrint()
         ),
         transports: [
-          new winston.transports.File({ filename: 'logs/ltijs_server.log',
+          new winston.transports.File({
+            filename: 'logs/ltijs_server.log',
             handleExceptions: true,
             json: true,
             maxsize: 250000, // 500kb (with two files)
             maxFiles: 1,
             colorize: false,
-            tailable: true })
+            tailable: true
+          })
         ],
         exitOnError: false
       })
@@ -139,106 +145,6 @@ class Provider {
     this.#ENCRYPTIONKEY = encryptionkey
     this.#server = new Server(options ? options.https : false, options ? options.ssl : false, this.#ENCRYPTIONKEY, loggerServer)
 
-    // Starts database connection
-    if (database.connection) {
-      if (!database.connection.useNewUrlParser) database.connection.useNewUrlParser = true
-      if (!database.connection.autoReconnect) database.connection.autoReconnect = true
-      if (!database.connection.keepAlive) database.connection.keepAlive = true
-      if (!database.connection.keepAliveInitialDelay) database.connection.keepAliveInitialDelay = 300000
-    } else {
-      database.connection = { useNewUrlParser: true, autoReconnect: true, keepAlive: true, keepAliveInitialDelay: 300000, connectTimeoutMS: 300000 }
-    }
-    this.#dbConnection.url = database.url
-    this.#dbConnection.options = database.connection
-
-    // Creating database schemas
-    const idTokenSchema = new Schema({
-      iss: String,
-      issuer_code: String,
-      user: String,
-      roles: [String],
-      userInfo: {
-        given_name: String,
-        family_name: String,
-        name: String,
-        email: String
-      },
-      platformInfo: {
-        family_code: String,
-        version: String,
-        name: String,
-        description: String
-      },
-      endpoint: {
-        scope: [String],
-        lineitems: String,
-        lineitem: String
-      },
-      namesRoles: {
-        context_memberships_url: String,
-        service_versions: [String]
-      },
-      createdAt: { type: Date, expires: 3600 * 24, default: Date.now }
-    })
-    const contextTokenSchema = new Schema({
-      path: String,
-      user: String,
-      context: { id: String, label: String, title: String, type: Array },
-      resource: { title: String, id: String }, // Activity that originated login
-      custom: { // Custom parameter sent by the platform
-        resource: String, // Id for a requested resource
-        system_setting_url: String,
-        context_setting_url: String,
-        link_setting_url: String
-      },
-      createdAt: { type: Date, expires: 3600 * 24, default: Date.now }
-    })
-    const platformSchema = new Schema({
-      platformName: String,
-      platformUrl: String,
-      clientId: String,
-      authEndpoint: String,
-      accesstokenEndpoint: String,
-      kid: String,
-      authConfig: {
-        method: String,
-        key: String
-      }
-    })
-    const keySchema = new Schema({
-      kid: String,
-      iv: String,
-      data: String
-    })
-    const accessTokenSchema = new Schema({
-      platformUrl: String,
-      iv: String,
-      data: String,
-      createdAt: { type: Date, expires: 3600, default: Date.now }
-    })
-    const nonceSchema = new Schema({
-      nonce: String,
-      createdAt: { type: Date, expires: 10, default: Date.now }
-    })
-
-    try {
-      mongoose.model('idToken', idTokenSchema)
-      mongoose.model('contextToken', contextTokenSchema)
-      mongoose.model('platform', platformSchema)
-      mongoose.model('privatekey', keySchema)
-      mongoose.model('publickey', keySchema)
-      mongoose.model('accesstoken', accessTokenSchema)
-      mongoose.model('nonce', nonceSchema)
-    } catch (err) {
-      if (this.#logger) this.#logger.log({ level: 'error', message: 'Message: ' + err.message + '\nStack: ' + err.stack })
-      provMainDebug('Model already registered. Continuing')
-    }
-
-    /**
-     * @description Database connection object.
-     */
-    this.db = mongoose.connection
-
     /**
      * @description Express server object.
      */
@@ -247,7 +153,7 @@ class Provider {
     /**
      * @description Grading service.
      */
-    this.Grade = new GradeService(this.getPlatform, this.#ENCRYPTIONKEY, this.#logger)
+    this.Grade = new GradeService(this.getPlatform, this.#ENCRYPTIONKEY, this.#logger, this.#Database)
 
     if (options && options.staticPath) this.#server.setStaticPath(options.staticPath)
 
@@ -255,67 +161,45 @@ class Provider {
     const sessionValidator = async (req, res, next) => {
       // Ckeck if request is attempting to initiate oidc login flow
       if (req.url === this.#loginUrl || req.url === this.#sessionTimeoutUrl || req.url === this.#invalidTokenUrl || this.#whitelistedUrls.indexOf(req.url) !== -1) return next()
-      if (req.url === this.#appUrl) {
+      if (req.url === this.#appUrl && !req.query.ltik) {
         let origin = req.get('origin')
         if (!origin || origin === 'null') origin = req.get('host')
         if (!origin) return res.redirect(this.#invalidTokenUrl)
         const iss = 'plat' + encodeURIComponent(Buffer.from(origin).toString('base64'))
-        return res.redirect(307, '/' + iss)
+
+        let token = {
+          issuer: iss,
+          path: this.#appUrl
+        }
+        // Signing context token
+        token = jwt.sign(token, this.#ENCRYPTIONKEY)
+
+        return res.redirect(307, this.#appUrl + '?ltik=' + token)
       }
 
       try {
-        let it = false
-        let urlArr = req.url.split('/')
-        let issuer = urlArr[1]
-        let path = ''
-        let isApiRequest = false
+        if (!req.query.ltik) return res.redirect(this.#invalidTokenUrl)
+        const validLtik = jwt.verify(req.query.ltik, this.#ENCRYPTIONKEY)
+        let user = false
+
+        const issuer = validLtik.issuer
+        const contextPath = _path.join(issuer, validLtik.path)
         const cookies = req.signedCookies
 
-        // Validate issuer_code to see if its a route or normal request
-        if (issuer.indexOf('plat') === -1) isApiRequest = true
-        if (!isApiRequest) {
-          try {
-            const decode = Buffer.from(decodeURIComponent(issuer.split('plat')[1]), 'base64').toString('ascii')
-            if (!validator.isURL(decode) && decode.indexOf('localhost') === -1) isApiRequest = true
-          } catch (err) {
-            provMainDebug(err.message)
-            if (this.#logger) this.#logger.log({ level: 'error', message: 'Message: ' + err.message + '\nStack: ' + err.stack })
-            isApiRequest = true
-          }
-        }
-
-        // Mount request path and issuer_code
-        if (isApiRequest) {
-          let requestParts
-          if (req.query.context) {
-            requestParts = req.query.context.split('/')
-          } else {
-            return res.status(400).send('Missing context parameter in request.')
-          }
-          issuer = encodeURIComponent(requestParts[1])
-          const _urlArr = []
-          for (const i in requestParts) _urlArr.push(requestParts[i])
-          urlArr = _urlArr
-
-          // Informs the system that is a API request
-          res.locals.isApiRequest = true
-        }
-        for (const i in urlArr) if (parseInt(i) !== 0 && parseInt(i) !== 1) path = path + '/' + urlArr[i]
-
-        // Mathes path to cookie
+        // Matches path to cookie
         for (const key of Object.keys(cookies)) {
           if (key === issuer) {
-            it = cookies[key]
+            user = cookies[key]
             break
           }
         }
 
         // Check if user already has session cookie stored in its browser
-        if (!it) {
+        if (!user) {
           provMainDebug('No cookie found')
           if (req.body.id_token) {
             provMainDebug('Received request containing token. Sending for validation')
-            const valid = await Auth.validateToken(req.body.id_token, this.getPlatform, this.#ENCRYPTIONKEY, this.#logger)
+            const valid = await Auth.validateToken(req.body.id_token, this.getPlatform, this.#ENCRYPTIONKEY, this.#logger, this.#Database)
             provAuthDebug('Successfully validated token!')
 
             // Mount platform token
@@ -343,61 +227,59 @@ class Provider {
             res.cookie(issuer, platformToken.user, this.#cookieOptions)
 
             // Store idToken in database
-            if (await Database.Delete('idToken', { issuer_code: issuer, user: valid.sub })) Database.Insert(false, 'idToken', platformToken)
+            if (await this.#Database.Delete('idtoken', { issuer_code: issuer, user: valid.sub })) this.#Database.Insert(false, 'idtoken', platformToken)
 
             // Mount context token
             const contextToken = {
-              path: issuer + '/',
+              path: contextPath,
               user: valid.sub,
               context: valid['https://purl.imsglobal.org/spec/lti/claim/context'],
               resource: valid['https://purl.imsglobal.org/spec/lti/claim/resource_link'],
               custom: valid['https://purl.imsglobal.org/spec/lti/claim/custom']
             }
 
-            res.cookie(issuer + '/', contextToken.resource.id, this.#cookieOptions)
+            res.cookie(contextPath, contextToken.resource.id, this.#cookieOptions)
 
             platformToken.platformContext = contextToken
 
             // Store contextToken in database
-            if (await Database.Delete('contextToken', { path: issuer + '/', user: valid.sub })) Database.Insert(false, 'contextToken', contextToken)
+            if (await this.#Database.Delete('contexttoken', { path: contextPath, user: valid.sub, 'resource.id': contextToken.resource.id })) this.#Database.Insert(false, 'contexttoken', contextToken)
 
+            res.locals.contextToken = req.query.ltik
             res.locals.token = platformToken
-            res.locals.login = true
 
             provMainDebug('Passing request to next handler')
             return next()
           } else {
             provMainDebug(req.body)
+            if (this.#logger) this.#logger.log({ level: 'error', message: req.body })
             provMainDebug('Passing request to session timeout handler')
             return res.redirect(this.#sessionTimeoutUrl)
           }
         } else {
           provAuthDebug('Cookie found')
-          res.cookie(issuer, it, this.#cookieOptions)
-          let valid = await Database.Get(false, 'idToken', { issuer_code: issuer, user: it })
-          if (!valid) return res.redirect(this.#invalidTokenUrl)
-          valid = valid[0]
+          // Gets correspondent id token from database
+          let idToken = await this.#Database.Get(false, 'idtoken', { issuer_code: issuer, user: user })
+          if (!idToken) throw new Error('No id token found')
+          idToken = idToken[0]
 
-          if (path) {
-            path = issuer + path
-            for (const key of Object.keys(cookies).sort((a, b) => b.length - a.length)) {
-              if (key === issuer) continue
-              if (path.indexOf(key) !== -1) {
-                let contextToken = await Database.Get(false, 'contextToken', { path: key, user: it, 'resource.id': cookies[key] })
-                if (!contextToken) throw new Error('No path cookie found')
-                contextToken = contextToken[0]
-                valid.platformContext = contextToken
-                break
-              }
+          let contextTokenName
+          // Matches path to cookie
+          for (const key of Object.keys(cookies).sort((a, b) => b.length - a.length)) {
+            if (contextPath.indexOf(key) !== -1) {
+              contextTokenName = key
+              break
             }
-          } else {
-            let contextToken = await Database.Get(false, 'contextToken', { path: issuer + '/', user: it, 'resource.id': cookies[issuer + '/'] })
-            if (!contextToken) throw new Error('No path cookie found')
-            contextToken = contextToken[0]
-            valid.platformContext = contextToken
           }
-          res.locals.token = valid
-          res.locals.login = false
+
+          // Gets correspondent context token from database
+          let contextToken = await this.#Database.Get(false, 'contexttoken', { path: contextTokenName, user: user, 'resource.id': cookies[contextTokenName] })
+          if (!contextToken) throw new Error('No context token found')
+          contextToken = contextToken[0]
+          idToken.platformContext = contextToken
+
+          res.locals.contextToken = req.query.ltik
+          res.locals.token = idToken
 
           provMainDebug('Passing request to next handler')
           return next()
@@ -405,7 +287,7 @@ class Provider {
       } catch (err) {
         provAuthDebug(err.message)
         if (this.#logger) this.#logger.log({ level: 'error', message: 'Message: ' + err.message + '\nStack: ' + err.stack })
-        provMainDebug('Error retrieving or validating token. Passing request to invalid token handler')
+        provMainDebug('Passing request to invalid token handler')
         return res.redirect(this.#invalidTokenUrl)
       }
     }
@@ -447,59 +329,42 @@ class Provider {
     })
 
     // Main app
-    this.app.post(this.#appUrl + ':iss', async (req, res, next) => {
-      if (!res.locals.isApiRequest) return this.#connectCallback(res.locals.token, req, res, next)
-      return next()
+    this.app.all(this.#appUrl, async (req, res, next) => {
+      return this.#connectCallback(res.locals.token, req, res, next)
     })
   }
 
   /**
      * @description Starts listening to a given port for LTI requests and opens connection to the database.
-     * @param {number} port - The port the Provider should listen to.
+     * @param {Object} [options] - Deployment options.
+     * @param {Number} [options.port] - Deployment port. 3000 by default.
+     * @param {Boolean} [options.silent] - If true, disables initial startup message.
      * @returns {Promise<true| false>}
      */
-  async deploy (port) {
+  async deploy (options) {
     provMainDebug('Attempting to connect to database')
-
     try {
-      this.db.on('connected', async () => {
-        provMainDebug('Database connected')
-      })
-      this.db.once('open', async () => {
-        provMainDebug('Database connection open')
-      })
-      this.db.on('error', async () => {
-        mongoose.disconnect()
-      })
-      this.db.on('reconnected', async () => {
-        provMainDebug('Database reconnected')
-      })
-      this.db.on('disconnected', async () => {
-        provMainDebug('Database disconnected')
-        provMainDebug('Attempting to reconnect')
-        setTimeout(async () => {
-          if (this.db.readyState === 0) {
-            try {
-              await mongoose.connect(this.#dbConnection.url, this.#dbConnection.options)
-            } catch (err) {
-              provMainDebug('Error in MongoDb connection: ' + err)
-              if (this.#logger) this.#logger.log({ level: 'error', message: 'Message: ' + err.message + '\nStack: ' + err.stack })
-            }
-          }
-        }, 1000)
-      })
-
-      if (this.db.readyState === 0) await mongoose.connect(this.#dbConnection.url, this.#dbConnection.options)
+      await this.#Database.setup()
     } catch (err) {
       provMainDebug(err.message)
       if (this.#logger) this.#logger.log({ level: 'error', message: 'Message: ' + err.message + '\nStack: ' + err.stack })
       return false
     }
+    const conf = {
+      port: 3000,
+      silent: false
+    }
 
-    /* In case no port is provided uses 3000 */
-    port = port || 3000
+    if (options && options.port) conf.port = options.port
+    if (options && options.silent) conf.silent = options.silent
     // Starts server on given port
-    this.#server.listen(port, 'LTI Provider is listening on port ' + port + '!\n\n LTI provider config: \n >Initiate login URL: ' + this.#loginUrl + '\n >App Url: ' + this.#appUrl + '\n >Session Timeout Url: ' + this.#sessionTimeoutUrl + '\n >Invalid Token Url: ' + this.#invalidTokenUrl)
+    this.#server.listen(conf, 'LTI Provider is listening on port ' + conf.port + '!\n\n LTI provider config: \n >Initiate login URL: ' + this.#loginUrl + '\n >App Url: ' + this.#appUrl + '\n >Session Timeout Url: ' + this.#sessionTimeoutUrl + '\n >Invalid Token Url: ' + this.#invalidTokenUrl)
+
+    // Sets up gracefull shutdown
+    process.on('SIGINT', async () => {
+      await this.close(options)
+      process.exit()
+    })
 
     return true
   }
@@ -508,11 +373,13 @@ class Provider {
    * @description Closes connection to database and stops server.
    * @returns {Promise<true | false>}
    */
-  async close () {
+  async close (options) {
     try {
+      if (!options || options.silent !== true) console.log('Closing server...')
       await this.#server.close()
-      this.db.removeAllListeners()
-      await this.db.close()
+      if (!options || options.silent !== true) console.log('Closing connection to the database...')
+      await this.#Database.Close()
+      if (!options || options.silent !== true) console.log('All done')
       return true
     } catch (err) {
       provMainDebug(err.message)
@@ -606,15 +473,13 @@ class Provider {
       const _platform = await this.getPlatform(platform.url)
 
       if (!_platform) {
-        const kid = await Auth.generateProviderKeyPair(this.#ENCRYPTIONKEY)
-        const plat = new Platform(platform.name, platform.url, platform.clientId, platform.authenticationEndpoint, platform.accesstokenEndpoint, kid, this.#ENCRYPTIONKEY, platform.authConfig, this.#logger)
+        const kid = await Auth.generateProviderKeyPair(this.#ENCRYPTIONKEY, this.#Database)
+        const plat = new Platform(platform.name, platform.url, platform.clientId, platform.authenticationEndpoint, platform.accesstokenEndpoint, kid, this.#ENCRYPTIONKEY, platform.authConfig, this.#logger, this.#Database)
 
         // Save platform to db
-        const isregisteredPlat = await Database.Get(false, 'platform', { platformUrl: platform.url })
-        if (!isregisteredPlat) {
-          provMainDebug('Registering new platform: ' + platform.url)
-          await Database.Insert(false, 'platform', { platformName: platform.name, platformUrl: platform.url, clientId: platform.clientId, authEndpoint: platform.authenticationEndpoint, accesstokenEndpoint: platform.accesstokenEndpoint, kid: kid, authConfig: platform.authConfig })
-        }
+        provMainDebug('Registering new platform: ' + platform.url)
+        await this.#Database.Insert(false, 'platform', { platformName: platform.name, platformUrl: platform.url, clientId: platform.clientId, authEndpoint: platform.authenticationEndpoint, accesstokenEndpoint: platform.accesstokenEndpoint, kid: kid, authConfig: platform.authConfig })
+
         return plat
       } else {
         return _platform
@@ -631,26 +496,22 @@ class Provider {
      * @param {String} url - Platform url.
      * @returns {Promise<Platform | false>}
      */
-  async getPlatform (url, ENCRYPTIONKEY, logger) {
+  async getPlatform (url, ENCRYPTIONKEY, logger, Database) {
     if (!url) throw new Error('No url provided')
     try {
-      const plat = await Database.Get(false, 'platform', { platformUrl: url })
+      const plat = Database !== undefined ? await Database.Get(false, 'platform', { platformUrl: url }) : await this.#Database.Get(false, 'platform', { platformUrl: url })
+
       if (!plat) return false
       const obj = plat[0]
 
       if (!obj) return false
-
-      let result
-      if (ENCRYPTIONKEY && logger !== undefined) {
-        result = new Platform(obj.platformName, obj.platformUrl, obj.clientId, obj.authEndpoint, obj.accesstokenEndpoint, obj.kid, ENCRYPTIONKEY, obj.authConfig, logger)
-      } else {
-        result = new Platform(obj.platformName, obj.platformUrl, obj.clientId, obj.authEndpoint, obj.accesstokenEndpoint, obj.kid, this.#ENCRYPTIONKEY, obj.authConfig, this.#logger)
-      }
+      const result = new Platform(obj.platformName, obj.platformUrl, obj.clientId, obj.authEndpoint, obj.accesstokenEndpoint, obj.kid, ENCRYPTIONKEY !== undefined ? ENCRYPTIONKEY : this.#ENCRYPTIONKEY, obj.authConfig, logger !== undefined ? logger : this.#logger, Database !== undefined ? Database : this.#Database)
 
       return result
     } catch (err) {
       provAuthDebug(err.message)
-      if (this.#logger) this.#logger.log({ level: 'error', message: 'Message: ' + err.message + '\nStack: ' + err.stack })
+      // If logger is set (Function was called by the Auth or Grade service) and is set to true, or if the scope logger variable is true, print the log
+      if ((logger !== undefined && logger) || this.#logger) logger !== undefined ? logger.log({ level: 'error', message: 'Message: ' + err.message + '\nStack: ' + err.stack }) : this.#logger.log({ level: 'error', message: 'Message: ' + err.message + '\nStack: ' + err.stack })
       return false
     }
   }
@@ -674,10 +535,10 @@ class Provider {
   async getAllPlatforms () {
     const returnArray = []
     try {
-      const platforms = await Database.Get(false, 'platform')
+      const platforms = await this.#Database.Get(false, 'platform')
 
       if (platforms) {
-        for (const obj of platforms) returnArray.push(new Platform(obj.platformName, obj.platformUrl, obj.clientId, obj.authEndpoint, obj.accesstokenEndpoint, obj.kid, this.#ENCRYPTIONKEY, obj.authConfig, this.#logger))
+        for (const obj of platforms) returnArray.push(new Platform(obj.platformName, obj.platformUrl, obj.clientId, obj.authEndpoint, obj.accesstokenEndpoint, obj.kid, this.#ENCRYPTIONKEY, obj.authConfig, this.#logger, this.#Database))
         return returnArray
       }
       return []
@@ -699,18 +560,35 @@ class Provider {
    */
   async redirect (res, path, options) {
     if (this.#whitelistedUrls.indexOf(path) !== -1) return res.redirect(path)
-    const newPath = res.locals.token.issuer_code + path
-    if (res.locals.login && (options && options.isNewResource)) {
+    const code = res.locals.token.issuer_code
+    const newPath = _path.join(code, path)
+
+    let token = {
+      issuer: code,
+      path: path
+    }
+    // Signing context token
+    token = jwt.sign(token, this.#ENCRYPTIONKEY)
+
+    if (options && options.isNewResource) {
       provMainDebug('Setting up path cookie for this resource with path: ' + path)
       res.cookie(newPath, res.locals.token.platformContext.resource.id, this.#cookieOptions)
-      res.locals.token.platformContext.path = newPath
-      if (await Database.Delete('contextToken', { path: newPath, user: res.locals.token.user })) Database.Insert(false, 'contextToken', res.locals.token.platformContext)
+
+      const newContextToken = {
+        resource: res.locals.token.platformContext.resource,
+        custom: res.locals.token.platformContext.custom,
+        context: res.locals.token.platformContext.context,
+        path: newPath,
+        user: res.locals.token.platformContext.user
+      }
+
+      if (await this.#Database.Delete('contexttoken', { path: newPath, user: res.locals.token.user, 'resource.id': res.locals.token.platformContext.resource.id })) this.#Database.Insert(false, 'contexttoken', newContextToken)
       if (options && options.ignoreRoot) {
-        Database.Delete('contextToken', { path: res.locals.token.issuer_code + '/', user: res.locals.token.user })
-        res.clearCookie(res.locals.token.issuer_code + '/')
+        this.#Database.Delete('contexttoken', { path: code + this.#appUrl, user: res.locals.token.user })
+        res.clearCookie(code + this.#appUrl)
       }
     }
-    return res.redirect(newPath)
+    return res.redirect(path + '?ltik=' + token)
   }
 }