+++ docs/changelog.md
@@ -8,6 +8,16 @@
 
 ### CHANGELOG
 
+#### V2.0.0
+
+> 2019-09-01
+>BREAKING CHANGES
+> - New authentication system, now uses a query parameter 'ltik', to pass the ltijs key to the application instead of embedding it into the path.
+> - As a result, simplified routing, without needing to account for the context path.
+> - Silent option added to the deploy method, that supresses the initial console logs.
+> - Ltijs now works with Postgresql database via the ltijs-postgresql plugin.
+> - Fixed staticPath option where it used to disable the invalidToken route if some index.html was present in the given path.
+> - Graceful shutdown added, closing connection to databases.
 
 #### V1.2.0